fixup balancing, poll LED
